ca_blazemeter_service/src/requirements.txt
cloudshell-shell-core>=2.3.0,<2.4.0
cloudshell-automation-api>=8.0,<8.3
requests